# [CPU] Optimize first inference latency for PagedAttention when running bf16