### Details:
 - *Use specific kernel for 2d f32 to bf16 conversion instead of multiple calls to cpu_convert*
     - there is an invocation of parallel_for inside cpu_convert, when copying count is small such as only a head size: 128, each core will only copy ~2 elements if core number is 60, this will result false sharing. The cost can reduce from ~1700ms to ~860ms after the fix. SDPA path will copy a block of heads such as 32*128, so it will not easily be impacted but very small prompt size should also suffer from the problem.
 - *Change the loop order from B,H,L to B,L,H due to the physical layout, can reduce the cost from ~860ms to ~830ms.*
 - *Changes in vLLM:  https://github.com/ilya-lavrenov/vllm/pull/15*

### Tickets:
 - *ticket-id*